feat: Refactor authentication middleware with new auth helpers and implement token sync restoration, alongside updates to enemy combat, game economy, and API routes.

+++ src/graphql/validation/schemas.ts
@@ -2,7 +2,7 @@ import { z } from 'zod';
 
 export const AddCoinsSchema = z.object({
   userId: z.string().min(1),
-  amount: z.number().int().positive().max(1000000, 'Amount too large'),
+  amount: z.number().positive().max(1000000, 'Amount too large'),
 });
 
 export const UseItemSchema = z.object({
